license correct

+++ src/chipMaskHDF5.cpp
@@ -23,7 +23,7 @@ void ChipMaskHDF5::openFile(){
     //}  
 }
 
-herr_t ChipMaskHDF5::writeDataSet(std::string chipID, slideRange& sliderange, unordered_map<uint64, Position1>& bpMap, int barcodeLen, int segment, int slidePitch, uint compressionLevel, int index){
+herr_t ChipMaskHDF5::writeDataSet(std::string chipID, slideRange& sliderange, unordered_map<uint64, Position1>& bpMap, uint32_t barcodeLen, uint8_t segment, uint32_t slidePitch, uint compressionLevel, int index){
     //generate dataSet space
     hid_t dataspaceID;
     hsize_t dims[RANK];
